added admin role and related features

+++ backend/src/controllers/cardController.ts
@@ -20,21 +20,34 @@ const findById = async (req: Request, res: Response) => {
 
 const create = async (req: Request, res: Response) => {
   const data = req.body;
-  const card = await cardService.create({ ...data, deckId: parseInt(req.params.deckId) });
-  res.status(201).json(card);
+  const id = parseInt(req.params.deckId);
+  if (id === req.user.id || req.user.role === "admin") {
+    const card = await cardService.create({ ...data, deckId: id });
+    res.status(201).json(card);
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 const update = async (req: Request, res: Response) => {
   const data = req.body;
   const id = parseInt(req.params.cardId);
-  const card = await cardService.update({ ...data, id });
-  res.status(200).json(card);
+  if (id === req.user.id || req.user.role === "admin") {
+    const card = await cardService.update({ ...data, id });
+    res.status(200).json(card);
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 const destroy = async (req: Request, res: Response) => {
   const id = parseInt(req.params.cardId);
-  await cardService.destroy(id);
-  res.status(200).json({ id });
+  if (id === req.user.id || req.user.role === "admin") {
+    await cardService.destroy(id);
+    res.status(200).json({ id });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 export default { getAllCards, getAllCardsByDeckId, findById, create, update, destroy };
